split gulp stuff into separate files
