+++ web/ui/gulpfile.js
@@ -1,221 +1,27 @@
-/* globals require: true, __dirname: true */
+/* jshint node: true */
 
-var gulp = require("gulp"),
-    gutil = require('gulp-util'),
-    concat = require("gulp-concat"),
-    rename = require("gulp-rename"),
-    uglify = require("gulp-uglify"),
-    jshint = require("gulp-jshint"),
-    sequence = require("run-sequence"),
-    babel = require("gulp-babel"),
-    sourcemaps = require("gulp-sourcemaps"),
-    karma = require('karma').server;
+let gulp = require("gulp");
 
-var paths = {
-    src: "src/",
-    srcBuild: "build/js/",
-    staticSrc: "static/",
-    staticBuild: "build/",
-    thirdpartySrc: "static/thirdparty/",
-    thirdpartyBuild: "static/thirdparty/"
-};
+// get all the gulp tasks
+require("./gulp/app.js");
+require("./gulp/thirdparty.js");
+require("./gulp/test.js");
 
-var babelConfig = {
-    presets: ["es2015"],
-};
+gulp.task("default", ["build"]);
 
-// files to be concatenated/minified to make
-// controlplane.js
-var controlplaneFiles = [
-    paths.src + "**/*.js"
-];
-
-var controlplanePartials = [
-    paths.src + "**/*.html"
-];
-
-// Third-party library files to be concatenated/minified to make thirdparty.js
-// NOTE: any changes here will not take effect until `release3rdparty` or `concat3rdparty`
-// task is run. Be sure to commit the new *MINIFIED* thirdparty.js file.
-var thirdpartyFiles = [
-    paths.thirdpartySrc + "jquery/jquery.js",
-    paths.thirdpartySrc + "jquery-timeago/jquery.timeago.js",
-    paths.thirdpartySrc + "jquery-ui/ui/jquery-ui.js",
-    paths.thirdpartySrc + "jquery-datetimepicker/jquery.datetimepicker.js",
-
-    paths.thirdpartySrc + "bootstrap/dist/js/bootstrap.js",
-    paths.thirdpartySrc + "bootstrap/js/tooltip.js",
-    paths.thirdpartySrc + "bootstrap/js/popover.js",
-
-    paths.thirdpartySrc + "elastic/elasticsearch.js",
-
-    paths.thirdpartySrc + "angular/angular.js",
-    paths.thirdpartySrc + "angular/angular-route.js",
-    paths.thirdpartySrc + "angular/angular-cookies.js",
-    paths.thirdpartySrc + "angular/angular-sanitize.js",
-    paths.thirdpartySrc + "angular-dragdrop/angular-dragdrop.js",
-    paths.thirdpartySrc + "angular-translate/angular-translate.js",
-    paths.thirdpartySrc + "angular-translate/angular-translate-loader-static-files/angular-translate-loader-static-files.js",
-    paths.thirdpartySrc + "angular-translate/angular-translate-loader-url/angular-translate-loader-url.js",
-    paths.thirdpartySrc + "angular-cache/angular-cache.js",
-    paths.thirdpartySrc + "angular-moment/angular-moment.js",
-    paths.thirdpartySrc + "angular-sticky/sticky.js",
-    paths.thirdpartySrc + "angular-location-update/angular-location-update.js",
-
-    paths.thirdpartySrc + "ng-table/ng-table.js",
-
-    paths.thirdpartySrc + "d3/d3.js",
-    paths.thirdpartySrc + "graphlib/graphlib.js",
-    paths.thirdpartySrc + "dagre-d3/dagre-d3.js",
-
-    paths.thirdpartySrc + "codemirror/lib/codemirror.js",
-    paths.thirdpartySrc + "codemirror/mode/properties/properties.js",
-    paths.thirdpartySrc + "codemirror/mode/yaml/yaml.js",
-    paths.thirdpartySrc + "codemirror/mode/xml/xml.js",
-    paths.thirdpartySrc + "codemirror/mode/shell/shell.js",
-    paths.thirdpartySrc + "codemirror/mode/javascript/javascript.js",
-    paths.thirdpartySrc + "angular-ui-codemirror/ui-codemirror.js",
-
-    paths.thirdpartySrc + "rison-node/rison.js",
-];
-
-// Enumerate the static assets (including thirdparty.js) so that the RPM/DEB
-//      packages only install what we really need
-var staticFiles = [
-    paths.staticSrc + '*.*',
-    paths.staticSrc + 'css/**/*.*',
-    paths.staticSrc + 'doc/**/*.*',
-    paths.staticSrc + 'fonts/**/*.*',
-    paths.staticSrc + 'help/**/*.*',
-    paths.staticSrc + 'i18n/**/*.*',
-    paths.staticSrc + 'ico/**/*.*',
-    paths.staticSrc + 'img/**/*.*',
-    paths.staticSrc + 'lib/bootstrap/dist/**/*.*',
-    paths.staticSrc + 'lib/codemirror/lib/*.css',
-    paths.staticSrc + 'lib/jquery-ui/themes/base/*.*',
-    paths.staticSrc + 'lib/jquery-datetimepicker/*.css',
-    paths.staticSrc + 'lib/thirdparty.*',
-    paths.staticSrc + 'scripts/**/*.*',
-    paths.staticSrc + 'lib/ng-table/ng-table.css'
-];
-
-gulp.task("default", ["concat", "copyStatic"]);
-gulp.task("release", function(){
-    // last arg is a callback function in case
-    // of an error.
-    sequence("lint", "concat", "uglify", "copyStatic", function(){});
-});
-
-// this needs to be run manually if 3rd party
-// code is updated, which should be infrequent
-gulp.task("release3rdparty", function(){
-    sequence("copyStatic", "concat3rdparty", "uglify3rdparty", function(){});
-});
-
-gulp.task("debug3rdparty", function(){
-    sequence("copyStatic", "concat3rdparty", "copyStatic", function(){});
-});
-
-gulp.task("concat", function(){
-    return gulp.src(controlplaneFiles)
-        .pipe(sourcemaps.init())
-            .pipe(babel(babelConfig))
-            .pipe(concat("controlplane.js"))
-        .pipe(sourcemaps.write("./", { sourceRoot: "src" }))
-        .pipe(gulp.dest(paths.srcBuild));
-});
-
-gulp.task("uglify", function(){
-    return gulp.src(paths.build + "controlplane.js")
-        .pipe(sourcemaps.init({loadMaps: true}))
-            .pipe(uglify())
-        .pipe(sourcemaps.write("./"))
-        .pipe(gulp.dest(paths.srcBuild));
-});
-
-gulp.task("concat3rdparty", function(){
-    return gulp.src(thirdpartyFiles)
-        .pipe(sourcemaps.init())
-            .pipe(concat("thirdparty.js"))
-        .pipe(sourcemaps.write("./", { sourceRoot: "thirdParty" }))
-        .pipe(gulp.dest(paths.thirdpartyBuild));
-});
-
-gulp.task("uglify3rdparty", function(){
-    return gulp.src(paths.thirdpartyBuild + "thirdparty.js")
-        .pipe(sourcemaps.init({loadMaps: true}))
-            .pipe(uglify())
-        .pipe(sourcemaps.write("./"))
-        .pipe(gulp.dest(paths.thirdpartyBuild));
-});
-
-gulp.task("copyStatic", function() {
-    gulp.src(staticFiles, {base: paths.staticSrc})
-        .pipe(gulp.dest(paths.staticBuild));
-
-    // gather partials from src
-    gulp.src(controlplanePartials)
-        .pipe(rename({dirname: ""}))
-        .pipe(gulp.dest(paths.staticBuild + "partials"));
-});
-
-gulp.task("watch", function(){
-    // concat js
-    gulp.watch(paths.src + "/**/*.js", ["concat"]);
-    // copy html templates
-    gulp.watch(paths.src + "/**/*.html", ["copyStatic"]);
-    // copy static content
-    gulp.watch(staticFiles, ["copyStatic"]);
-    // copy translations
-    gulp.watch(paths.staticSrc + "/i18n/*", ["copyStatic"]);
-    // TODO - preprocess CSS
-});
-
-//
-// The equivalent manual execution of karma is:
-//  ./node_modules/karma/bin/karma start karma.conf.js --single-run \
-//      --log-level debug \\
-//      --browsers PhantomJS \\
-//      --reporters progress,junit,coverage
-gulp.task('test', function (done) {
-  karma.start({
-    configFile: __dirname + '/karma.conf.js',
-    singleRun: true,
-    logLevel: "debug",
-    browsers: ["PhantomJS"],
-    reporters: ["progress","junit","coverage","threshold"],
-  }, function(exitStatus) {
-    // Workaround for 'formatError' based on suggestions from
-    //   http://stackoverflow.com/questions/26614738/issue-running-karma-task-from-gulp
-    // but tweaked to use (apparently new) PluginError
-    var err = exitStatus ? new gutil.PluginError('test', 'There are failing unit tests') : undefined;
-    done(err);
-  });
-});
-
-
-// brings up web browser and auto-runs tests as they
-// are saved and edited
-gulp.task('tdd', function (done) {
-  karma.start({
-    configFile: __dirname + '/karma.conf.js',
-    browsers: ["Chrome"],
-    reporters: ["html"],
-    autoWatch: true
-  }, function(exitStatus) {
-    // Workaround for 'formatError' based on suggestions from
-    //   http://stackoverflow.com/questions/26614738/issue-running-karma-task-from-gulp
-    // but tweaked to use (apparently new) PluginError
-    var err = exitStatus ? new gutil.PluginError('test', 'There are failing unit tests') : undefined;
-    done(err);
-  });
-});
-
-
-gulp.task("lint", function(){
-    return gulp.src(controlplaneFiles)
-        .pipe(jshint(".jshintrc"))
-        .pipe(jshint.reporter("jshint-stylish"))
-        .pipe(jshint.reporter("fail"));
-});
-
+/*
+ * you probably want to do one of these:
+ *
+ * `gulp`
+ * lints and builds the js library
+ *
+ * `gulp watch`
+ * watches the filesystem and continuously builds the js lib
+ *
+ * `gulp test`
+ * runs unit tests
+ *
+ * `gulp tdd`
+ * watches the filesystem and continuously builds and runs
+ * the unit tests
+ */
